requirements.txt
connexion
connexion[swagger-ui]
python_dateutil
setuptools
gunicorn 
Flask
pymongo
dnspython
python-dotenv
werkzeug
google-api-python-client
python-bsonjs